增加SprDebugNode
【详细描述】N/A
【修改类型】Optimize
【问题链接】N/A
【项目关联】Default

+++ UtilModules/EpollEventHandler/PPipe.h
@@ -27,8 +27,8 @@
 class PPipe : public IEpollEvent
 {
 public:
-    PPipe(int fd, std::function<void(int, std::string, void*)> cb = nullptr, void *arg = nullptr);
-    PPipe(const std::string& fileName, std::function<void(int, std::string, void*)> cb = nullptr, void* arg = nullptr);
+    PPipe(int fd, std::function<void(ssize_t, std::string, void*)> cb = nullptr, void *arg = nullptr);
+    PPipe(const std::string& fileName, std::function<void(ssize_t, std::string, void*)> cb = nullptr, void* arg = nullptr);
     virtual ~PPipe();
 
     void* EpollEvent(int fd, EpollType eType, void* arg) override;
